requirements.txt
streamlit
python-pptx
google-generativeai
openai
anthropic
